# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puraani Jeans APIs",
        "description": "Puraani Jeans API documentation",
        "contact": {
            "email": "developer@puranijeans.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": ""
    },
    "servers": [
        {
            "url": "http://puranijeans.test/",
            "description": "dev"
        }
    ],
    "paths": {
...
        "/api/filter/closet/{slug}/product": {
            "post": {
                "tags": [
                    "Closet"
                ],
                "summary": "Store Products",
                "operationId": "getFilteredClosetProducts",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "store slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Resend OTP",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "filters": {
                                        "description": "Network Id of customer's phone number",
                                        "type": "object",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        },
        "/api/closet/{slug}/product": {
            "get": {
                "tags": [
                    "Closet"
                ],
                "summary": "Store Products",
                "operationId": "getClosetProducts",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "store slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        },
...
        "/api/closet/{slug}/category/{catSlug}/product": {
            "get": {
                "tags": [
                    "Closet"
                ],
                "summary": "Manage Closet Category Products",
                "operationId": "getClosetCategoryProducts",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "catSlug",
                        "in": "path",
                        "description": "category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        },
...
        "/api/product/{productId}": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Product detail",
                "operationId": "getProductDetail",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Get product detail",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "referrer_type": {
                                        "description": "referrer type",
                                        "type": "string",
                                        "example": "Featured Product"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "user_access_token": {
                "type": "apiKey",
                "description": "user access token",
                "name": "Authorization",
                "in": "header"
            },
            "client_id": {
                "type": "apiKey",
                "description": "client id",
                "name": "x-client-id",
                "in": "header"
            },
            "client_secret": {
                "type": "apiKey",
                "description": "client secret",
                "name": "x-client-token",
                "in": "header"
            }
        }
    }
}